+++ blocks/form/generateForm.js
@@ -25,6 +25,7 @@ function createLabel(field, fieldData, placeholders) {
   return label;
 }
 
+/*
 function setCommonAttributes(field, fieldData, placeholders, index) {
   field.id = fieldData.id;
   field.name = fieldData.name;
@@ -33,9 +34,9 @@ function setCommonAttributes(field, fieldData, placeholders, index) {
     || fieldData['example-placeholder']
     || fieldData.name
     || '';
-  field.value = fieldData['default-value'] || '';
   field.dataset.index = index;
 }
+*/
 
 const createSelect = async (fieldData) => {
   const select = document.createElement('select');
@@ -98,12 +99,10 @@ const createToggle = (fieldData) => {
   return toggleSwitch;
 };
 
-const createCheckbox = (fieldData, placeholders) => {
+const createCheckbox = (fieldData) => {
   const field = createInput(fieldData);
   if (fieldData['default-value']) field.checked = true;
-  field.value = placeholders[toCamelCase(fieldData['label-placeholder'])]
-    || fieldData['label-placeholder']
-    || fieldData.name;
+  field.value = 'yes';
 
   return field;
 };
@@ -192,7 +191,7 @@ async function createField(fieldData, form, placeholders, index) {
   return field;
 }
 
-export default async function createForm(formConfigPath, onSubmit) {
+export default async function createForm(formConfigPath) {
   const [resp, placeholders] = await Promise.all([
     fetch(formConfigPath),
     getPlaceholder(null),
@@ -230,10 +229,6 @@ export default async function createForm(formConfigPath, onSubmit) {
       placeholders.formButtonSubmit || 'Send'
     }</button>`,
   );
-
-  if (onSubmit) {
-    form.onsubmit = onSubmit;
-  }
 
   return form;
 }
